Random Password Generator with CLI tool
About the Tool

This generates a password for your services like Facebook, Instagram, etc., copy it to the clipboard, and stores them in a text file named passwords.txt .

To view the saved passwords Click Here

Features

This tool generates an user customized password. User can enter the Length, number of lower and upper cases characters, number of numerals and special characters.
After password generation it copies it to the clipboard, from where it can pe easily pasted anywhere.
It uses the Fisher-Yates shuffle which runs in O(n) time and also proven to be a perfect shuffle, so it takes less time with high randomness.